+++ include/grid.h
@@ -3,28 +3,28 @@
 #include "object.h"
 #include "shape.h"
 
-vector<shared_ptr<Shape2D>> build(const vec2 &_pos, const vec2 &_bmin,
-                                  const vec2 &_bmax);
+vector<shared_ptr<Shape2D>> build(const dvec2 &_pos, const dvec2 &_bmin,
+                                  const dvec2 &_bmax);
 
 class Grid2D : public Object2D {
   friend class VTKWriter;
 
 private:
-  vec2 cornerMin, cornerMax;
+  dvec2 cornerMin, cornerMax;
   int maxX, maxY;
-  float dx, dy;
-  vector<float> data;
-  function<void(Ray2D &, float, float &)> cellAction;
+  double dx, dy;
+  vector<double> data;
+  function<void(Ray2D &, double, double &)> cellAction;
   function<void(Ray2D &, const IntersectResult2D&)> hitAction;
-  function<float(float)> refractiveIndexFunction;
+  function<double(double)> refractiveIndexFunction;
 
 public:
-  Grid2D(const vec2 &_pos, const vec2 &_bmin, const vec2 &_bmax, int _maxX,
-         int _maxY, function<void(Ray2D &, float, float &)> _cellAction,
+  Grid2D(const dvec2 &_pos, const dvec2 &_bmin, const dvec2 &_bmax, int _maxX,
+         int _maxY, function<void(Ray2D &, double, double &)> _cellAction,
          function<void(Ray2D &, const IntersectResult2D&)> _hitAction,
-         function<float(float)> _refractiveIndexFunction)
+         function<double(double)> _refractiveIndexFunction)
       : Object2D(build(_pos, _bmin, _bmax), 0, _pos), maxX(_maxX), maxY(_maxY),
-        data(maxX * maxY, 0.0f), cellAction(_cellAction), hitAction(_hitAction),
+        data(maxX * maxY, 0.0), cellAction(_cellAction), hitAction(_hitAction),
         refractiveIndexFunction(_refractiveIndexFunction) {
     cornerMin = root->box->aabb.bmin;
     cornerMax = root->box->aabb.bmax;
@@ -35,10 +35,10 @@ public:
   void action(Ray2D &ray, const IntersectResult2D &result,
               vector<Ray2D> &createdRays);
 
-  float sum();
-  float avg();
-  float var();
-  float stddev();
+  double sum();
+  double avg();
+  double var();
+  double stddev();
 
   void reset();
