+++ src/mirror.cpp
@@ -1,17 +1,17 @@
 #include "mirror.h"
 
-vector<shared_ptr<Shape2D>> build(const vec2 &_pos, const vec2 &_opticalAxis,
-                                  const function<vec2(float)> _shapeFunction,
+vector<shared_ptr<Shape2D>> build(const dvec2 &_pos, const dvec2 &_opticalAxis,
+                                  const function<dvec2(double)> _shapeFunction,
                                   int _segments) {
   vector<shared_ptr<Shape2D>> lines;
   for (int i = 0; i < _segments; ++i) {
-    vec2 ap = _shapeFunction(i * (1.0f / _segments));
-    vec2 bp = _shapeFunction((i + 1) * (1.0f / _segments));
+    dvec2 ap = _shapeFunction(i * (1.0 / _segments));
+    dvec2 bp = _shapeFunction((i + 1) * (1.0 / _segments));
 
-    vec2 am = {-ap.x, ap.y};
-    vec2 bm = {-bp.x, bp.y};
+    dvec2 am = {-ap.x, ap.y};
+    dvec2 bm = {-bp.x, bp.y};
 
-    float rotationAngle = orientedAngle({0.0f, 1.0f}, _opticalAxis);
+    double rotationAngle = orientedAngle({0.0, 1.0}, _opticalAxis);
 
     ap = rotate(ap, rotationAngle);
     bp = rotate(bp, rotationAngle);
